refactor(api): rename /api/1-asset-transactions route to /api/asset-transactions
The `1-` prefix on the API route folder was an accidental carry-over from
numbering the docs spec folder (docs/specs/asset-transactions →
docs/specs/1-asset-transactions); it leaked into the actual HTTP path
(/api/1-asset-transactions). Rename the route folder and every reference to it:
the 5 client fetch URLs in assetTransactionService.ts, the route handlers' log
labels/doc comments, the route tests' imports + request URLs, and the CLAUDE.md/
AGENTS.md route-path mentions. The `docs/specs/1-asset-transactions/*` spec
citations are the spec's real path and are left untouched.

No behavior change (the app calls its own API; no external consumers, no
firestore.rules coupling). tsc clean; full suite 69 files / 1215 tests green.
